0.4.0 Release

+++ src/swganh/messages/update_transform_message.h
@@ -42,9 +42,9 @@ struct UpdateTransformMessage : public swganh::messages::BaseSwgMessage<UpdateTr
     
     void onSerialize(anh::ByteBuffer& buffer) const {
         buffer.write(object_id);
-        buffer.write<int16_t>(position.x * 4.0f + 0.5f);
-        buffer.write<int16_t>(position.y * 4.0f + 0.5f);
-        buffer.write<int16_t>(position.z * 4.0f + 0.5f);
+        buffer.write<int16_t>(static_cast<int16_t>(position.x * 4.0f + 0.5f));
+        buffer.write<int16_t>(static_cast<int16_t>(position.y * 4.0f + 0.5f));
+        buffer.write<int16_t>(static_cast<int16_t>(position.z * 4.0f + 0.5f));
         buffer.write(update_counter);
         buffer.write(static_cast<uint8_t>(0));
         buffer.write(heading);
